Add files via upload
For Calcium Imaging; variables in place to obtain numbers and percentages of neurons in a field of view and cellular responses to various agonists on a per-experiment basis through an iterative function. Function requires which agonists were added in what time interval and the processed (normalized to background and to a value of 1) fluorescence values. Script also contains commands to provide ggplot of response runs.